refactor(#2162): extract ScanProgress metering into scan_progress module
Keeps the over-threshold producer.rs growth minimal (campsite rule, epic
#1116): the new incremental-emission machinery lives in its own small module.
4.3: telemetry-only (no user-facing API/config per non-goals), no doctrine change.
